requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.37.4
botocore==1.37.4
Brotli==1.0.9
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dashscope==1.22.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.10.0
fastjsonschema==2.21.1
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
interchange==2021.0.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.39
langchain-text-splitters==0.3.6
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.2
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.28.1
networkx==3.4.2
numpy==1.26.4
openai==1.63.0
openpyxl==3.1.5
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
pansi==2024.11.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py2neo==2021.2.4
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
yarg==0.1.9
yarl==1.18.3
